debugger: Improve keys command
Use the new evalInFrame() function to improve the keys command to handle
exceptions raised while printing the expression's keys. Print the result
directly, rather than calling into the debuggee to print it.

Also handle the cases gracefully where the argument isn't an object, or
doesn't have any own properties.

Change 'keys' to 'own properties' in the description, in case the user is
looking up help for the keys command because they don't know what keys
are...