value: Explicit cast from g(u)int64 to (u)int64_t
This is necessary on platforms where the custom GLib types and the C
stdlib types are the same width but not aliased to the same type. This
happens for historical reasons on MacPorts and Homebrew, where int64_t is
long long, while gint64 is long. Unfortunately this cannot be changed, as
per https://docs.gtk.org/glib/types.html#gint64 (see
https://gitlab.gnome.org/GNOME/glib/-/merge_requests/2841 for more
details; changing the typedef is an ABI break because the types are
reflected in C++ mangled names.)

Sadly, I don't know of a way to catch this in CI unless we gained a CI
runner for a platform that has this weirdness. So we will probably have to
continue to fix it on a case-by-case basis when it occurs.

Closes: #608