build: Use DYLD_LIBRARY_PATH instead of DYLD_FALLBACK_LIBRARY_PATH
I had previously thought that this was sufficient on macOS; if a library
isn't found in the default path, we fall back to the build directory.
However, this causes hard-to-track-down errors when running tests while
there is also an earlier version of GJS installed to the default library
path.

So, better to just have the same behaviour as on other Unix.