modules/console: Do not avoid pretty printing undefined values
The GJS console does not currently print results if they are undefined.

This is because of an isUndefined early return in gjs_console_eval_and_print that prevent calls to the pretty print func.

Remove the early return for parity with other JS runtimes.

Fixes https://gitlab.gnome.org/GNOME/gjs/-/issues/621