importer: do not iterate search path dir if error exists
If g_file_enumerate_children encounters an error for a search path it will return NULL. This can happen for directories that do not exist.

It is a programmer error to pass NULL to g_file_enumerator_iterate, so break out of the directory processing loop if we do not have a directory to iterate on.