prettyPrint: handle more cases in nested objects
GJS logs warnings if we pretty print an object containing null as a value. This is because a WeakSet cannot take in null values but null has typeof object.

GJS also does not print [GjsFileImporter root] objects if it is nested within another object, such as for imports.gi. This is because the logic on whether to call toString differs between prettyPrint and formatObject.

Guard explicitly against null and apply the same logic to formatObject as were in prettyPrint.