build: Build Cairo from subproject if not found
This makes Cairo a required dependency, but it will be built from a
subproject if it is not available in the system.

There are a few reasons to do this: it makes the build system code much
simpler, it removes a build configuration that is rarely tested, and it
will be simpler to handle the transition to gobject-introspection-tests
(see https://gitlab.gnome.org/GNOME/gobject-introspection/-/issues/489)
if Cairo can be a required dependency there, too.