js: Move UniquePtr into external ArrayBuffers
ArrayBuffers now take a unique pointer instead of a raw pointer. For
JS::NewExternalArrayBuffer(), the free callback is part of
JS::BufferContentsDeleter. For JS::NewArrayBufferWithContents, the free
function was always JS_free(); in our _gjs_memdup2() case, it was already
incorrect to use that function, so use JS::NewExternalArrayBuffer()
instead.