modules: add source map support
Using the mozilla source-map library, hook into module resolve, async resolve and load to extract the source map URL from the JS source code and build the SourceMapConsumer. When the URL is not a data uri, it is resolved relative to the location of the source file.

Populating the source map is always graceful and should/will not fail given bad inputs.