js: Heap can no longer be stored in WeakCache
We used to be able to store JS::Heap<T> in JS::WeakCache, but that is no
longer possible because JS::GCPolicy<JS::Heap> does not implement
needsSweep().

Add a class Gjs::WeakPtr<T> that behaves the same as JS::Heap<T> but has
the appropriately defined GCPolicy to be able to be stored in WeakCache.