module: handle import query component
GJS currently fails to import modules if specifiers contain query parameters.

This is because the query question mark delimiter has been %-encoded when canonicalizing or resolving relative uris . GJS then attempts to load the resource or file with this encoded uri and fails.

Ensure the special "?" delimiter is not mangled and stop using the query component when loading resources or files. However, use the full uri in the module registry to enable cache-busting behaviour.

Fixes https://gitlab.gnome.org/GNOME/gjs/-/issues/618