byteArray: Copy GBytes data in ByteArray.fromGBytes()
We previously used the GBytes' data pointer as the backing store for the
ArrayBuffer. This is wrong for two reasons. One, the data pointer may not
meet the alignment requirements of ArrayBuffer. Two, the data pointer may
point to immutable memory. Just always copy the data and have the
ArrayBuffer own it.