byteArray: Copy data directly into ArrayBuffer
In gjs_byte_array_from_data(), previously we copied the data and then gave
that copy to the array buffer as an external pointer. It's much simpler to
allocate the ArrayBuffer with the right amount of storage and then copy
the data directly into the ArrayBuffer's storage.

Rename the function to gjs_byte_array_from_data_copy() to make it clearer
that it copies the data.