arg: Un-templateize gjs_arg_unset()
We can use a simple memset() to zero out the whole union. There is a
comment in gjs_gi_argument_init_default() about type-punning a libffi
buffer as GIArgument but we don't seem to do this anymore.

Some testing with Godbolt indicates that both the templated and
untemplated versions are compiled to a single XOR instruction with -O2.