console: persist REPL history to user defined path
We already use history.h and add_history. Allow a new env variable GJS_REPL_HISTORY which contains the user defined path of the history file. If this env var is "" then no history is persisted. If this env var is not set a default location under XDG cache is used.

Add new property in GjsContext repl-history and set it in console main() based on env var. If this property is set in main() or by program embedders to non-NULL then read_history on GjsContext construct and write_history to that path whenever a new command is entered. Also warn the user if read/write_history is unsuccessful.

Fixes https://gitlab.gnome.org/GNOME/gjs/-/issues/645