tests: Test for current behaviour of skipped parameters
I had written tests for what should have been the behaviour of skipped
parameters all along, and marked them pending. However, this desired
behaviour is an API break and it's not clear when we'll be able to
implement it, if ever. So it's more useful to have tests for the current
behaviour to make sure that API is not broken accidentally.