arg-cache: NotIntrospectable should not assume GITypeInfo is available
One would think that a non-introspectable parameter would always have
introspection info available (otherwise, how would you know it existed and
was not introspectable). But because of a missing API to get GITypeInfo
for an instance parameter (GNOME/gobject-introspection#334), we do have to
deal with the case where GITypeInfo is not available. If that happens, we
should not crash.

This makes the NotIntrospectable arg cache inherit SkipAll instead of
GenericIn, which basically has the effect of removing its m_type_info
member.

See: #655