arg-cache: Remove parameter pack expansion in ArgsCache
The parameter pack expansion made things difficult to debug in this file,
because it produces confusing error messages at compile time (the compiler
errors point to the `new T(args...)` in Argument::make(), and you have to
find the actual mismatched arguments by trial and error.)

Instead of creating the T inside Argument::make(), create it at the call
site, pass it in, and let the template parameter be inferred.

This is slightly more verbose but not actually all that much more, and
it's much easier to debug.