arg-cache: Use Maybe for return types
Functions that do not return a value may have a null GITypeInfo for their
return value, and a GITypeTag of GI_TYPE_TAG_VOID. Especially the latter
can cause confusion when we forget to check is_pointer() of the
GITypeInfo. Use Maybe to indicate clearly when the it actually means there
is no return value.