auto: Add function to safely move JS::UniqueChars to Gjs::AutoChar
This is only possible because currently, js_free() and g_free() both
ultimately call free() to free the data. Doing this would cause crashes
if SpiderMonkey were to stop supporting embedders using the system
allocator in the future. In that case, this function would have to copy
the string. That's why we are isolating this usage now.

See for context:
https://gitlab.gnome.org/GNOME/gjs/-/merge_requests/866#note_2027399
https://gitlab.gnome.org/GNOME/gjs/-/issues/79