arg-cache: Check for GdkAtom first
In some cases, g_registered_type_info_get_g_type() would return
G_TYPE_NONE for GdkAtom, if the type initialization function couldn't yet
be called because the Gdk dynamic library was not loaded. This would cause
the is_gdk_atom() check not to be reached, because the condition of
GI_IS_STRUCT_INFO && G_TYPE_NONE && !is_gtype_struct would be reached
first.

This had no effect because the result was the same, but we should not have
this dead code, and would like to add custom handling for GdkAtom in the
future.