arg-types-inl: Replace static_type_name() with MarshallingInfo struct
This is a first step towards eliminating the <T, TAG> templates, which
have proved to often cause bugs on non-x64 architectures due to breaking
assumptions about what is typedefed to what. We will put more information
in these MarshallingInfo structs.

This commit breaks static_type_name() for gboolean and GType, but the
following commits will restore that.