object: Handle mismatch in type of property with type of C accessor
There can be edge cases (from broken introspection data, see
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/1708
for an example) where the type of the property doesn't match the type of
the C setter or getter.

In that case, log a warning (so that these broken APIs can be found and
fixed) and fall back to the slow g_object_get_property() path.

Note that it is not sufficient to compare the GITypeInfos with
g_base_info_equal(), because for non-basic types, e.g. the GdkWindow* type
of gtk_widget_get_window(), the GIR file reports the type of the
GtkWidget:window property as GdkWindow, not GdkWindow*.
Also, integers of specific widths are stored as g_param_spec_int or
g_param_spec_long when they are GObject properties. So we write a custom
equality function.

Closes: #677