arg-types-inl: Replace <T, TAG> pairs with a single TAG
The TAG was necessary to specify what to do with types that were typedefed
to each other. For example, gboolean is int32_t, but must be treated like
bool, so we added GI_TYPE_TAG_BOOLEAN to the template specialization: e.g.
<gboolean, GI_TYPE_TAG_BOOLEAN>.

This approach had several problems. The main one is that because typedefs
can be different between 32-bit and 64-bit, T and TAG would become
unmatched and fail to compile on 32-bit, which would go unnoticed because
we don't have a CI pipeline for 32-bit.

Any information that is needed for each TAG, we add to MarshallingInfo.
That is, the corresponding GITypeTag, the real type corresponding to the
TAG (accessible via Tag::RealT<TAG>), and the strict and relaxed JSValue
holder types. (Relaxed is renamed to JSValuePackType because I find that
clearer: it is the type corresponding to the way the value is packed
inside JSValue in SpiderMonkey.)

In most of the cases, we add a specialization for those types where TAG is
already the same type as Tag::RealT<TAG>, so that the template type can be
inferred. Except in the case of js_value_to_c(), where the previous code
mostly ignored the TAG -- but we actually don't want that anymore!