gi: Move away from GIInfoType where possible
GIInfoType is being removed in gobject-introspection 2.0, and type
checking macros will be used instead.

(Most of) the macros exist already in gobject-introspection 1.0, although
they check the GIInfoType field rather than the GType. We didn't use them
previously, because it was usually easier to switch on the GIInfoType.
However, adopting them now will give us confidence that the new-style code
is correct and make the port to g-i 2.0 easier to review.