overrides: add simpler GObject.Value constructor API
GObject.Value in GJS is currently very C-like and requires calling init with a GType before calling a type specific setter.

Add a new JS override to the constructor which allows users to pass in the type and initial value to reduce this to a single line.

Fixes https://gitlab.gnome.org/GNOME/gjs/-/issues/456