Gtk: Builder constructor overrides in GTK4
Overrides Gtk.Builder in GTK4 only, with a subclass that accepts construct
parameters analogous to new_from_file(), new_from_resource(), and
new_from_string().

Also accepts a "callbacks" construct parameter in which case a scope is
automatically created.

The adjustment in ObjectInstance::typecheck_impl() is needed because
apparently otherwise we can't call builder.get_current_object() in the
vfunc implementation - the vfunc gets a GtkBuilder* C pointer that has the
GType of GtkJSBuilder, but because it's typed as GtkBuilder it doesn't
have the JS subclass info available.

Based on prototype by Sonny Piers <sonnyp@gnome.org>