module loader: Throw on resolution fail instead of returning nullable URI
A number of functions, such as resolveSpecifier(), return a nullable
module or URI, only to throw an ImportError later. Instead, throw the
ImportError.

In ModuleLoader we have to override InternalModuleLoader.resolveSpecifier
and additionally resolve bare specifiers to internal modules if the super
call fails.

This is preparation for import maps. Adapted from Evan's refactors.

Co-Authored-By: Evan Welsh <contact@evanwelsh.com>