module loader: Make resolveRelativeResourceOrFile() return Uri object
Use the new gjs_uri_object() to return a Uri object, which is not nullable
(the nullability of the previous string return value was maybe a mistake?)

This makes InternalModuleLoader.resolveRelativePath() trivial, so we drop
it and inline it instead. This actually saves a call to parseURI() in
source map population.

Co-Authored-By: Evan Welsh <contact@evanwelsh.com>