arg-cache: Make array fixed size a Maybe<size_t>
In girepository 2.0, size returned by gi_type_info_array_fixed_size() will
be a size_t, and -1 will not be used as a sentinel value. Adapt the API in
gi/info.h to work like this already and make all the necessary changes at
call sites.

Note that in ArgCache::FixedSizeArray we will not store all 64 bits of the
fixed size on 64-bit systems, because that bloats the argument caches. We
will limit fixed size arrays to UINT32_MAX which is already an improvement
over girepository 1.0's INT_MAX anyway.