arg-cache: Make destroy notify and user data index Maybe<unsigned>
Similar to the previous commit, in girepository 2.0, the index returned by
gi_arg_info_get_closure_index() and gi_arg_info_get_destroy_index() will
be an unsigned int, and -1 will not be used as a sentinel value. Adapt the
API in gi/info.h to work like this already and make all the necessary
changes at call sites.