object: Remove dead code in resolve_no_info()
According to the documentation for G_TYPE_IS_CLASSED, interface GTypes are
not classed. So this loop through the interfaces would always skip every
interface. It's not clear whether this has always been the case, but it
may have been broken since being introduced in 74171bf3.

It doesn't seem to be necessary, because any GObject class implementing a
GInterface must override all the properties from the interface. (If you
try to skip doing that, you'll get a critical in class_init.)

The second dead code block seems like it was necessary in the past, but no
longer now that we resolve on interface prototypes.

Neither of these lazy_define_gobject_property calls are ever hit in any
part of the test suite, according to code coverage.