object: Change m_closures back to vector
std::unordered_set is 32 bytes bigger than std::vector so adding it would
increase the size of ObjectInstance which could easily add several hundred
kb of memory usage to gnome-shell. It seems like we only need random
access when disconnecting a signal or receiving a done notification from a
scope-notify callback, neither of which are hot paths.

So revert that part of the change and turn invalidate_closure_vector()
into a template that can handle either unordered_set (for m_vfuncs) or
vector (for m_closures).