maint: Switch to flat eslint config
eslint 8.x is no longer supported, and newer versions deprecated
both the .json and .yml config formats in favor of flat configs
exported from eslint.config.js files.

Port to the new config format, so we can easily update to supported
eslint versions in the future.

For the actual rules, switch to the shared `eslint-config-gnome` package
in World/javascript, which is a port of the existing rules to the new
config system in a shareable format (already used by gnome-shell and
others).