function: Use StackArgInfo for closure invalidate notifier
This code was not covered by tests, so add a test based on the reproducer
from the issue.

The problem was trying to take ownership from StackArgInfo to AutoArgInfo.
The refcount field of a stack info is marked with INVALID_REFCOUNT, so
that should not work.

Add a deleted constructor converting StackArgInfo and StackTypeInfo to
OwnedInfo, which would make this bug actually fail to compile. Also
clarify the situation in comments.

Also add move constructors to StackInfo to make this explicit, and delete
the copy constructors. (Copy constructors could be added later if needed,
but this seems easier for now because in g-i 2.0 we will have to manually
clear StackInfos on destruction.)

Closes: #691