error: Make Error.isError work on GErrors
In browsers, Error.isError() returns true for DOMExceptions, even though
they are not part of the JS language. Analogously, we should do the same
for GErrors.

Note the situation described in #700, there are two kinds of GErrors. This
does not yet work for the unusual kind (returned from a function.)