README.MSVC.md: Update Visual Studio build info
Let people know with SpiderMonkey-140.x, it is only currently supported
to build GJS with clang-cl, as there are constructs and syntax that is
only supported with GCC or CLang both in the code and the SpiderMonkey
headers.

Also note that for the rest of the stack, it is now recommended to build
the dependencies with Visual Studio 2019 instead of Visual Studio 2015.