CI: Install IWYU, rather than building it
Fedora includes an IWYU package for some time now. So we don't have to
build it ourselves when creating the Docker image.

Since that makes the Dockerfile for debug images pretty much identical to
that for non-debug images, just use the same Dockerfile for both, and
delete the debug Dockerfile.