gjs/deprecation: Use a larger deprecation stack for JS invoked requests
When using warnDeprecatedOncePerCallsite() from JS we need to use a
larger stack to track the call site, otherwise we may end up considering
the same call site for different actual callers.

In fact when using this in Gio.js override the call site is always the
same, but we care about where that function is being invoked from.