Gtk: Allow access by name on array returned from Gtk.Builder.get_objects
This allows accessing the objects from get_object as additional named
properties on the array:

  const {obj1, obj2} = builder.get_objects();

Unfortunately we have to use a Proxy here instead of just defining the
named properties, as it's not possible to get the list of valid object
IDs. You can get a partial list by iterating the array and calling
gtk_buildable_get_id() on each object, but the objects aren't guaranteed
to implement GtkBuildable.

Note the Proxy will keep the builder alive!