Require GLib 2.86.0
In order to reduce the likelihood of conflicts with untested combinations
of different module versions, require GLib 2.86.0 going forward, and
remove the compatibility code for GioUnix. (We still keep the
compatibility code in Ns::resolve() for GLibUnix. This is a slightly odd
situation, as we now have entirely different code paths to handle GLibUnix
and GioUnix, but on the other hand, GLibUnix is a lot simpler to handle.)

This also allows removing gjs_lookup_gtype() which was a wrapper around
GI::Repository::find_by_gtype() with GioUnix compatibility code.

Also the clang-analyzer-friendly g_assertion_message has been in GLib for
quite some time, we can remove the workaround for that as well.