arg: Validate UTF-8 being passed to JS when not directly returned
This was fixed for return values and out parameters in !1019 but not when
the value was returned indirectly, for example as an array element.
Discovered because there was still a test marked pending on this closed
issue.

In order to avoid a memory leak in the error case, release the array in
BasicTypeContainer::out. But this will still need to be handled in other
cases, see #716.

See: #658