tests: Update pending tests
If a pending test was blocked on a closed issue or merge request that's
now merged, we can run the test. If it was blocked on a merge request that
was closed without being merged, we can delete the test. Finally, if it
now still doesn't work, update the pending message to point to the correct
issue.