overrides/Gio: Provide meaningful stacktrace on async methods failures
In case an async DBus method implementation is also an async function
and it throws for whatever reason we fail with a generic promise failure
that is not helpful to understand where the problem was.

An example is:

  (gnome-shell:1686483): Gjs-WARNING **: 11:50:49.493: Unhandled promise
    rejection.
    To suppress this warning, add an error handler to your promise chain
    with .catch() or a try-catch block around your await expression.
  Stack trace of the failed promise:
    NotifyAsync@resource:///org/gnome/shell/ui/notificationDaemon.js:158:22
    _handleMethodCall@resource:///org/gnome/gjs/modules/core/overrides/Gio.js:374:35
    _wrapJSObject/<@resource:///org/gnome/gjs/modules/core/overrides/Gio.js:409:34
    @resource:///org/gnome/shell/ui/init.js:21:20

So in case the async method returns a promise, catch, log and return the
error to the DBus client.

In fact so far no error was really returned to the client and any
failing client request was never fulfilled.