overrides/Gio: Do not return errors to invocations if already returned
In case GDBus servers methods implemented asynchronously fail for
wahtever reason we were returning an error to the client.

However, this is not needed nor safe to do in case the invocation had
returned *before* that the error was thrown.

In fact Gio.DBusMethodInvocation return functions take the ownership of
the invocation itself (no matter the invocation happens or not) and
calling the invocation multiple times, may lead to crashes.

Given that the API does not provide a way to detect this, let's wrap the
affected functions with a safer function that prevents returning an instance
multiple times and that tells us how to behave in case of error.
Tracking all this only for the lifecycle of the instances themselves.

Now we can then have this other kind of case in which:
 - A method invocation returned properly to the client
 - The server code failed still failed while doing it