maint: Use IWYU associated pragma
This allows marking a header file as being 'associated' with a .cpp file
just like the .h file of the same name automatically is. This allows
removing several special cases that we had in tools/run_iwyu.sh.

Apparently 'association' allows IWYU to make better recommendations than
the --check_also option did, so we adjust some of the includes as well.