maint: Remove usage of NULL from C++, preferring nullptr
In C++, nullptr is more typesafe than NULL. There may be edge cases when
using NULL that trigger an integer overload of a function rather than a
pointer overload. C code remains unchanged, because it does not have a
nullptr keyword. We were already doing this in all new code, but this
changes all existing instances.

In some cases, other solutions are preferable than just replacing NULL
with nullptr. For example, making use of default arguments.

Also review all uses of "NULL" in comments, and change to "nullptr"
(specific) or "null" (generic) as appropriate.