maint: Remove usage of gboolean, TRUE, and FALSE where possible
Note that gboolean is an integer type, and TRUE and FALSE are integers.
We only keep these when the gboolean is significant for a function
signature (e.g. GSourceFunc) or we explicitly need to pass an integer in
order to select the proper template or overload.

Also remove TRUE and FALSE from comments unless we are specifically
talking about the integer constants.