maint: Remove usage of gsize and gssize where possible
We only keep gsize and gssize when passing them as an out parameter (since
GLib may define them to be a different width than size_t and ssize_t on
some platforms) and when the type itself is significant for selecting the
proper template or overload. These cases are marked with comments.