gjs-gdbus-wrapper: Clarify ownership of method invocation
As per the recent changes in the Gio override, we're *always* calling a
GDBusMethodInvocation return method on `::handle-method-call` handlers,
so we should not drop any reference here, but rather consider it gone the
invocation is passed to the JS side.

However, this is not something we can do in gjs, since any transfer-full
method would just copy the value we're passing to, so we are not really
dropping last reference without actually doing it explicitly.

So let's clarify the invocation ownership with code or comments