gjs-gdbus-wrapper: Pass the invocation ownership to JS
Since the method invocation ownership is completely handled by JS side,
we should consider also its' lifecycle to be handled there, and we
should not mess with reference counting at C level, given that an
invocation may be around for longer than C expects, and so unreferencing
it too early may be wrong.

So just leave all this being handled by the GC instead.

Note that technically here we could have just called the invocation with
`g_dbus_method_invocation_return_error` if nothing handles the
notification, but it's not something we can do in C anyways.