maint: Use consistent variable names for GjsContext*
Consistently call it "gjs_context", or "self" if in a GObject method of
GjsContext. In GjsCoverage and GjsProfiler specifically, we can refer to
it as "coverage_context" or "profiling_context".

(Not "gjs", that is GjsContextPrivate*, and "cx" which is JSContext*, and
not "context" or "js_context" which could be understood to be JSContext*.)