maint: Reformat all function declarations and headers
This makes formatting of all function headers consistent with the way
clang-format does it. (And gets rid of that annoying quirk of clang-format
where an annotation ends up cluttering the same line as the function name
when it's a template specialization.)

Previously the code was a mix of clang-format style, GLib-style, and some
other quirky style.

As for annotations, they go on the same line only when the entire
declaration fits on one line, or the entire function body.