function: Remove unnecessary inline lambda
It doesn't capture anything, better to replace it with a static method
since that will be clearer in debugging.

Currently shows up in stack traces like this:

  GjsCallbackTrampoline::callback_closure
  GjsCallbackTrampoline::create_closure()::$_0::operator()(ffi_cif*,
    void*, void**, void*) const
  GjsCallbackTrampoline::create_closure()::$_0::__invoke(ffi_cif*, void*,
    void**, void*)
  ffi_closure_unix64_inner()

It is also, theoretically, slightly more efficient at runtime, since the
lambda technically creates an instance of an anonymous class for the
lambda. (But the compiler may inline the lambda anyway with
optimization.)