CI: Run cpplint on the whole codebase
After 8 years with a "ratchet" rule where the number of cpplint warnings
can only go down and not up, the number of warnings is small enough that
we can just fix them all at once. In future, we'll fail the CI job if
there are any new warnings. So we can remove the "ratchet".

Add tools/run_cpplint.sh to run the same analysis locally that is run in
CI.