gi: Allow optional inout arguments to be null
When checking whether a provided argument can be null, we currently
only consider the `nullable` annotation.

While that is correct for "in" parameters, for "inout" parameters
it is the `optional` annotation that indicates that the argument
itself can be null (like for "out" parameters), while `nullable`
means that the argument may *point* to null.

It should be valid to pass `null` for optional "inout" parameters
(like it is in C), so explicitly handle this case.