maint: Don't use C89-style declarations
This is not C and it's not 1989 :-) No need to declare a variable and then
initialize it on a separate line. Move declarations inside for-loop init
blocks where possible, and onto the same line as their initialization
otherwise.

This makes a style consistent that we have been using for quite a while in
all new code.