toggle: Add assertion to idle handler
self->m_toggle_handler cannot be null here, because the only way this
function can be reached is if enqueue() was called, requiring a handler.
But clang-analyzer cannot detect this, so add an assertion.

This is probably a sign that the toggle queue needs refactoring. It is odd
that you have to provide a handler on every call to enqueue(), but it must
be the same handler every time.