maint: Add infrastructure for clang-tidy
Adds a CI job that runs clang-tidy on any changed files, as well as a
manual job that runs clang-tidy on the whole codebase. Also includes a
tools/run_clang_tidy.sh script for reproducing the results of the CI job
locally.

In this commit, no checks are enabled. In the following commits, checks
will be enabled one by one.