jsapi-util: Fix regression in RSS behaviour
In bd9aa0eb (!1051) I misread the existing code as "It claimed to lower
the trigger threshold if the RSS shrank below 75% of the threshold, but it
actually raised the threshold." This wasn't correct, it did lower the
threshold to 25% above the _current_ value. Restore the previous
behaviour and clarify it in a comment.