info: Fix return type of GI::CallableInfo::closure_native_address()
See https://gitlab.gnome.org/GNOME/glib/-/issues/3859; the void** return
type is a mistake.

Change the return type of GI::CallableInfo::closure_native_address() in
gi/info.h accordingly.

This function is used in GjsCallbackTrampoline::closure() which is renamed
to get_func_ptr() as I think the overuse of the word "closure" for FFI
closures, function pointers, etc., contributed to the confusion.

On that note, also rename the variables at call sites of get_func_ptr() to
better reflect what they represent: CallbackIn::in()'s "closure" becomes
"callback" since it is the callback pointer passed to a C function; and
ObjectInstance::hook_up_vfunc_impl()'s "method_ptr" becomes
"vfunc_slot_ptr" and we cast it to void**, because it is a pointer to the
slot in the GObject class structure where the vfunc function pointer is
stored.