maint: Check for suspicious sizeof expressions
I had previously disabled this check in the clangd configuration because
it triggered on g_clear_pointer(), but that seems to have been fixed in a
subsequent version of GLib. It still does trigger a false positive on
g_set_object(), but we only have a couple of uses of that.