maint: Check for use-after-move
This fixes all warnings flagged by the 'bugprone-use-after-move' check. We
also list g_steal_pointer() as a "std::move-like" function, although I
don't know for sure if that will work since it takes a location and not an
object.
https://clang.llvm.org/extra/clang-tidy/checks/bugprone/use-after-move.html

High severity. There is one test where we do it intentionally.