maint: Check for unused nontrivial variables
The compiler already warns about unused trivial variables. But until now
we didn't get any warnings about, for example, rooting a JS value
unnecessarily. This adds the 'bugprone-unused-local-non-trivial-variable'
check and makes it check every type of variable with an exclude list. (The
default is to check just a few standard library types.)
https://clang.llvm.org/extra/clang-tidy/checks/bugprone/unused-local-non-trivial-variable.html