maint: Remove usage of typedef
In non-C code, we want to define type aliases with 'using' rather than
'typedef', which reads weird and is less versatile. Clang-tidy's
'modernize-use-using' check can flag instances of this.

In the case of the C-specific pattern of "typedef (struct/enum)
_UnderscoredName { ... } RealName;", replace typedef with simply a struct
or enum declaration, not 'using'. In C++, the struct or enum name is just
put into the surrounding namespace, unlike in C, so it doesn't require a
type alias.