boxed: Simplify StructPrototype and UnionPrototype initialization
This refactor makes initialization of BoxedPrototype infallible, and
removes the need to store and GC-trace a property key for the default
constructor; we just need to store the index of the constructor's
GI::FunctionInfo to be able to retrieve it on demand. The fallible part
(converting GI::FunctionInfo::name to a JS property key) moves to the
constructor body.

It also removes the unsafe API of having to check
has_zero_args_constructor() and has_default_constructor() before getting
info about the constructors. The info is now encapsulated in a Maybe.