maint: Use all modernize clang-tidy checks except a few
Rather than enabling all the rest of the modernize checks one by one, we
can enable all of them at once except a few. (Nothing currently in the
codebase triggers any of them.)

As for the exceptions, avoiding C arrays will be difficult since we have
to pass them to C functions; and trailing return types and
<cstddef>-style headers are a stylistic choice that I don't think we
should make right now.

We need to exclude config.h from the checks as well, or else
'modernize-macro-to-enum' triggers on GJS_VERSION.